- [ZBX-820] fixed updating downstream item applications when upstrim application is updated without changing applications (Vedmak)
git-svn-id: svn://svn.zabbix.com/trunk@8038 97f52cf1-0a1b-0410-bd0e-c28be96e8082

+++ frontends/php/include/items.inc.php
@@ -479,6 +479,9 @@
 		'snmpv3_securityname','snmpv3_securitylevel','snmpv3_authpassphrase','snmpv3_privpassphrase',
 		'formula','trends','logtimefmt','valuemapid','delay_flex','params','ipmi_sensor','applications','templateid');
 //*/
+//sdii($item['applications']);
+		$upd_app = !is_null($item['applications']);
+// sdi($upd_app);		
 		$item_in_params = $item;
 
 		$item_data = get_item_by_itemid_limited($itemid);
@@ -547,13 +550,20 @@
 
 			$child_item_params['hostid'] = $db_tmp_item['hostid'];
 			$child_item_params['templateid'] = $itemid;
-			$child_item_params['applications'] = get_same_applications_for_host($item['applications'], $db_tmp_item['hostid']);
+			
+			if($upd_app){
+				$child_item_params['applications'] = get_same_applications_for_host($item['applications'], $db_tmp_item['hostid']);
+			}
+			else{
+				$child_item_params['applications'] = null;
+			}
 
 			if(!check_db_fields($db_tmp_item, $child_item_params)){
 				error('Incorrect arguments pasted to function [update_item]');
 				return false;
 			}
 
+			
 			$result = update_item($db_tmp_item['itemid'], $child_item_params);		// recursion!!!
 
 			if(!$result)
@@ -571,11 +581,14 @@
 		$item_old = get_item_by_itemid($itemid);
 		DBexecute('UPDATE items SET lastlogsize=0 WHERE itemid='.$itemid.' AND key_<>'.zbx_dbstr($item['key_']));
 
-		if(isset($_REQUEST['applications_visible'])){
-			$result = DBexecute('DELETE FROM items_applications WHERE itemid='.$itemid);
-			foreach($item['applications'] as $appid){
-				$itemappid=get_dbid('items_applications','itemappid');
-				DBexecute('INSERT INTO items_applications (itemappid,itemid,applicationid) VALUES ('.$itemappid.','.$itemid.','.$appid.')');
+
+		if($upd_app){
+			if(isset($_REQUEST['applications_visible'])){
+				$result = DBexecute('DELETE FROM items_applications WHERE itemid='.$itemid);
+				foreach($item['applications'] as $appid){
+					$itemappid=get_dbid('items_applications','itemappid');
+					DBexecute('INSERT INTO items_applications (itemappid,itemid,applicationid) VALUES ('.$itemappid.','.$itemid.','.$appid.')');
+				}
 			}
 		}
 
@@ -689,7 +702,6 @@
 
 		if($item_data['type'] == ITEM_TYPE_HTTPTEST)
 			$item['applications'] = get_applications_by_itemid($itemid);
-
 	return update_item($itemid,$item);
 	}
